Refactor

+++ src/commonMain/kotlin/io/github/katerinapetrova/mpplib/Base64.kt
@@ -5,7 +5,7 @@ interface Base64Encoder {
     fun encodeToString(src: ByteArray): String {
         val encoded = encode(src)
         return buildString(encoded.size) {
-            encoded.forEach { append(it.toChar()) }
+            encoded.forEach { append(it.toInt().toChar()) }
         }
     }
 }
